+++ front/.storybook/preview.ts
@@ -1,3 +1,5 @@
+import "../src/index.css";
+
 import type { Preview } from "@storybook/react-vite";
 
 const preview: Preview = {
