+++ front/src/stories/components/ui/Button.stories.tsx
@@ -1,0 +1,25 @@
+import type { Meta, StoryObj } from "@storybook/react-vite";
+
+import { Button } from "@/components/ui/button";
+
+const meta = {
+  title: "Components/UI/Button",
+  component: Button,
+  tags: ["autodocs"],
+  args: {
+    children: "Save",
+  },
+} satisfies Meta<typeof Button>;
+
+export default meta;
+
+type Story = StoryObj<typeof meta>;
+
+export const Default: Story = {};
+
+export const Outline: Story = {
+  args: {
+    variant: "outline",
+    children: "Cancel",
+  },
+};
